Created a sorted orderbook subclass.
